RB info
A command to query resource attributes. At the moment, this only
supports review requests, and is not able to list available attributes.
(If there is a sane way to do this with the current state of the API, I
have not been able to determine what it is.)

Args: rbt info <resource_selector> <attribute>
Options: --request (resource id)